linter and react set up
